+++ src/device/component/param/AlgParseHelper.cpp
@@ -2,10 +2,12 @@
 // Licensed under the MIT License.
 
 #include "AlgParseHelper.hpp"
+#include <cstddef>
+
 namespace libobsensor {
 std::vector<OBCameraParam_Internal_V0> AlgParseHelper::alignCalibParamParse(uint8_t *data, uint32_t size) {
     std::vector<OBCameraParam_Internal_V0> output;
-    for(int i = 0; i < static_cast<int>(size / D2C_PARAMS_ITEM_SIZE); i++) {
+    for(size_t i = 0; i < size / D2C_PARAMS_ITEM_SIZE; i++) {
         output.push_back(*(OBCameraParam_Internal_V0 *)(data + i * D2C_PARAMS_ITEM_SIZE));
     }
     return output;
